# api/openapi.yaml
openapi: 3.1.0
info:
  title: Cleaning App API
  version: 0.1.0
  description: |
    一人暮らし向け掃除管理アプリのAPI。
    契約ファースト（ADR-003）の唯一の正本。このファイルからクライアント・サーバースタブを生成する。
servers:
  - url: http://localhost:8080
    description: ローカル開発

# MVPは認証なし（X-User-Idヘッダで識別）。V2でKeycloak/JWTに移行予定
security: []
...
components:
  parameters:
...
    UserIdHeader:
      name: X-User-Id
      in: header
      required: true
      description: 初回起動時に発行したユーザーUUID
      schema:
        type: string
        format: uuid
  schemas:
...
    Furniture:
      type: object
      required: [id, roomId, name, gridX, gridY, gridW, gridH, rotation, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        roomId:
          type: string
          format: uuid
          description: 所属する部屋のID
        name:
          type: string
          example: ソファ
        presetKey:
          anyOf:
            - type: string
            - type: 'null'
          description: プリセット由来なら識別キー。自由名称で追加した場合はnull
          example: sofa
        gridX:
          type: integer
          description: 部屋内の相対セルX（部屋左上が原点）
        gridY:
          type: integer
          description: 部屋内の相対セルY
        gridW:
          type: integer
          description: 幅（セル数）。回転後の占有幅（rotation が 90/270 なら素材の高さと入れ替わる）
        gridH:
          type: integer
          description: 高さ（セル数）。回転後の占有高さ（rotation が 90/270 なら素材の幅と入れ替わる）
        rotation:
          type: integer
          enum: [0, 90, 180, 270]
          default: 0
          description: 時計回りの回転角（度）。描画のみに影響し、占有矩形は常に gridW×gridH で軸平行
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    # 家具の「作成入力」
    FurnitureCreate:
      type: object
      required: [name, gridX, gridY, gridW, gridH]
      properties:
        name:
          type: string
        presetKey:
          anyOf:
            - type: string
            - type: 'null'
          description: プリセット由来なら識別キー。自由名称の場合は省略またはnull
        gridX:
          type: integer
        gridY:
          type: integer
        gridW:
          type: integer
        gridH:
          type: integer
        rotation:
          type: integer
          enum: [0, 90, 180, 270]
          default: 0
          description: 時計回りの回転角（度）。省略時は 0

    # 家具の「更新入力」。省略したフィールドは変更しない（部分更新）
    FurnitureUpdate:
      type: object
      description: 指定したフィールドのみ更新する。すべて省略可能だが、少なくとも1つ指定すること
      minProperties: 1
      properties:
        name:
          type: string
        gridX:
          type: integer
          description: 部屋内の相対セルX
        gridY:
          type: integer
          description: 部屋内の相対セルY
        gridW:
          type: integer
          description: 幅（セル数）
        gridH:
          type: integer
          description: 高さ（セル数）
        rotation:
          type: integer
          enum: [0, 90, 180, 270]
          description: 時計回りの回転角（度）

    # 掃除単位。cleaning-record featureと共有するエンティティ
    # 間取りエディタが生成し（部屋種別によりseed）、cleaning-recordが掃除状態を更新する
...
paths:
...
  /rooms/{roomId}/furniture:
    post:
      summary: 部屋に家具を追加する
      description: |
        指定した部屋内にグリッド座標を指定して家具を追加する。
        プリセットキーを指定するとプリセット由来の家具として記録される。
        座標はroomId の部屋内の相対セル位置で指定する。
      operationId: createFurniture
      parameters:
        - $ref: '#/components/parameters/UserIdHeader'
        - name: roomId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FurnitureCreate'
      responses:
        '201':
          description: 作成された家具
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Furniture'
        '400':
          description: 入力が不正
        '404':
          description: 部屋が存在しない、または他ユーザーの部屋
  /furniture/{furnitureId}:
    patch:
      summary: 家具を更新する
      description: |
        家具の名称・グリッド座標・サイズを部分更新する。
        指定しなかったフィールドは変更されない。
      operationId: updateFurniture
      parameters:
        - $ref: '#/components/parameters/UserIdHeader'
        - name: furnitureId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FurnitureUpdate'
      responses:
        '200':
          description: 更新後の家具
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Furniture'
        '400':
          description: 入力が不正
        '404':
          description: 家具が存在しない、または他ユーザーの家具

    delete:
      summary: 家具を削除する
      description: |
        家具と、配下のパーツをすべて削除する。
        連鎖削除はDBのON DELETE CASCADEで行う。
      operationId: deleteFurniture
      parameters:
        - $ref: '#/components/parameters/UserIdHeader'
        - name: furnitureId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: 削除成功
        '404':
          description: 家具が存在しない、または他ユーザーの家具
...